refactor(frontend): split PlaygroundConfigSection
Follows the parked Phase-1 plan: the file keeps its path (barrel import
unchanged) and its parts move into the co-located PlaygroundConfigSection/
dir, in a one-directional import DAG to avoid the circular-dep trap:

  types <- constants <- {metadata, llmConfig, schemaData} <- atoms <- adapters

The component body's two contiguous blocks become hooks called at their
original positions, so hook order, count and dep arrays are unchanged:

  useModelConfigurePopover   the configure popover (26-entry useMemo dep
                             array kept as a memo, not made a component)
  useFieldSlots              the per-field slot renderers

The four public types moved to types.ts and are re-exported from
PlaygroundConfigSection.tsx, so no importer changed and no shim was left.

PlaygroundConfigSection.tsx: 2003 -> 478 lines.
