+++ web/packages/agenta-entity-ui/src/DrillInView/components/PlaygroundConfigSection.tsx
@@ -10,47 +10,22 @@
  * - Supports custom moleculeAdapter for specialized behavior
  * - Schema drives control selection via getSchemaAtPath
  * - Model config popover injected via fieldHeader slot
+ *
+ * Composition root only: the adapter, the pure helpers, the configure popover and
+ * the drill-in field slots live in `./PlaygroundConfigSection/*`.
  */
 
 import {memo, useMemo, useCallback, useEffect, useRef, useState} from "react"
 
-import {
-    type EntitySchema,
-    type EntitySchemaProperty,
-    getSchemaAtPath as getSchemaAtPathUtil,
-} from "@agenta/entities/shared"
 import {workflowMolecule} from "@agenta/entities/workflow"
-import type {Workflow} from "@agenta/entities/workflow"
 import type {DataPath} from "@agenta/shared/utils"
-import {getOptionsFromSchema, getValueAtPath, setValueAtPath} from "@agenta/shared/utils"
-import {HeightCollapse} from "@agenta/ui"
-import type {
-    FieldActionsSlotProps,
-    FieldContentSlotProps,
-    FieldHeaderSlotProps,
-    MoleculeDrillInAdapter,
-} from "@agenta/ui/drill-in"
+import type {MoleculeDrillInAdapter} from "@agenta/ui/drill-in"
 import {useDrillInUI} from "@agenta/ui/drill-in"
-import {formatLabel} from "@agenta/ui/drill-in"
 import {SharedEditor} from "@agenta/ui/shared-editor"
-import {ArrowLeft, CaretDown, CaretRight, MagicWand} from "@phosphor-icons/react"
-import {Button, Dropdown, Popover, Tabs, Tooltip, Typography} from "antd"
 import clsx from "clsx"
-import deepEqual from "fast-deep-equal"
-import type {Atom, WritableAtom} from "jotai"
-import {atom} from "jotai"
 import {useAtom, useAtomValue, useSetAtom} from "jotai"
-import {selectAtom} from "jotai/utils"
 import yaml from "js-yaml"
 
-import {
-    getModelSchema,
-    getLLMConfigValue,
-    getLLMConfigProperties,
-    HookConfigControl,
-    CodeConfigControl,
-    SchemasConfigControl,
-} from "../SchemaControls"
 import {feedbackConfigModeAtomFamily} from "../SchemaControls/FeedbackConfigurationControl"
 import {
     validateConfigAgainstSchema,
@@ -58,605 +33,28 @@ import {
 } from "../SchemaControls/schemaValidator"
 
 import {MoleculeDrillInView} from "./MoleculeDrillInView"
-import {FallbackConfigTab} from "./PlaygroundConfigSection/FallbackConfigTab"
-import {ModelConfigEditor} from "./PlaygroundConfigSection/ModelConfigEditor"
-import {RetryConfigTab} from "./PlaygroundConfigSection/RetryConfigTab"
+import {defaultAdapter} from "./PlaygroundConfigSection/adapters"
+import {SIBLING_GROUP_KEYS, type SiblingGroupKey} from "./PlaygroundConfigSection/constants"
+import {
+    collectAgentaMetadata,
+    reattachAgentaMetadata,
+    stripAgentaMetadata,
+} from "./PlaygroundConfigSection/metadata"
+import {
+    hasRenderableConfigSections,
+    isSiblingGroupKey,
+    pickActiveData,
+} from "./PlaygroundConfigSection/schemaData"
+import type {ConfigViewMode, PlaygroundConfigSectionProps} from "./PlaygroundConfigSection/types"
+import {useFieldSlots} from "./PlaygroundConfigSection/useFieldSlots"
+import {useModelConfigurePopover} from "./PlaygroundConfigSection/useModelConfigurePopover"
 
-// ============================================================================
-// TYPES
-// ============================================================================
-
-interface SchemaQueryResult {
-    isPending: boolean
-    isError: boolean
-    error: Error | null
-    data: {agConfigSchema?: PathSchema | null} | null
-}
-
-type PathSchema = EntitySchema | EntitySchemaProperty
-type ConfigureTabKey = "model" | "fallback" | "retry"
-interface FallbackDetailState {
-    mode: "new" | "edit"
-    index: number | null
-    draft: Record<string, unknown>
-}
-
-function isEntitySchema(value: unknown): value is PathSchema {
-    if (!value || typeof value !== "object") return false
-    const schemaType = (value as Record<string, unknown>).type
-    return typeof schemaType === "string"
-}
-
-/**
- * Adapter interface for the data source that PlaygroundConfigSection reads from.
- * Defaults to workflowMolecule when not provided.
- */
-export interface ConfigSectionMoleculeAdapter {
-    atoms: {
-        /** Entity data with `.parameters` — used for hasParameters checks and popover data */
-        data: (id: string) => Atom<{parameters?: Record<string, unknown>} | null>
-        /** Base data (pre-draft) with `.parameters` */
-        serverData: (id: string) => Atom<{parameters?: Record<string, unknown>} | null>
-        /** Draft data */
-        draft: (id: string) => Atom<unknown>
-        /** Whether entity has local changes */
-        isDirty: (id: string) => Atom<boolean>
-        /** Schema query state */
-        schemaQuery: (id: string) => Atom<SchemaQueryResult>
-        /** ag_config schema */
-        agConfigSchema: (id: string) => Atom<PathSchema | null>
-    }
-    reducers: {
-        update: WritableAtom<unknown, [id: string, changes: Record<string, unknown>], void>
-        discard: WritableAtom<unknown, [id: string], void>
-    }
-    drillIn: {
-        getRootData?: (data: unknown) => unknown
-        getChangesFromRoot?: (data: unknown, rootData: unknown, path: DataPath) => unknown
-        getValueAtPath?: (data: unknown, path: DataPath) => unknown
-        getRootItems?: (data: unknown) => unknown[]
-        getChangesFromPath?: (data: unknown, path: DataPath, value: unknown) => unknown
-        valueMode?: "native" | "structured"
-    }
-    selectors: {
-        schemaAtPath: (params: {id: string; path: (string | number)[]}) => Atom<unknown>
-    }
-    /** @deprecated Use useSetAtom(mol.reducers.update) instead */
-    set?: {
-        update: (id: string, changes: Record<string, unknown>) => void
-    }
-}
-
-function hasParameters(data: {parameters?: Record<string, unknown>} | null | undefined): boolean {
-    return Boolean(data?.parameters && Object.keys(data.parameters).length > 0)
-}
-
-const FALLBACK_POLICY_OPTIONS = ["off", "availability", "capacity", "access", "context", "any"].map(
-    (value) => ({label: value, value}),
-)
-const RETRY_POLICY_OPTIONS = ["off", "availability", "capacity", "transient", "any"].map(
-    (value) => ({label: value, value}),
-)
-const DEFAULT_RETRY_CONFIG = {
-    max_retries: null as number | null,
-    base_delay: null as number | null,
-}
-const PROMPT_EXTENSION_KEYS = [
-    "fallback_configs",
-    "fallback_policy",
-    "retry_config",
-    "retry_policy",
-]
-
-const createFallbackConfigKey = () => {
-    if (typeof crypto !== "undefined" && typeof crypto.randomUUID === "function") {
-        return crypto.randomUUID()
-    }
-
-    return `${Date.now()}-${Math.random()}`
-}
-
-const updateConfigKey = (
-    base: Record<string, unknown> | undefined,
-    key: string,
-    value: unknown,
-) => {
-    const next = {...(base ?? {})}
-    if (value === null || value === undefined) {
-        delete next[key]
-    } else {
-        next[key] = value
-    }
-    return next
-}
-
-const getResettableLLMConfigKeys = (llmConfigProps: Record<string, unknown>) =>
-    Object.keys(llmConfigProps).filter(
-        (key) => key !== "model" && !PROMPT_EXTENSION_KEYS.includes(key),
-    )
-
-const resetLLMParameterFields = ({
-    base,
-    resetBase,
-    resetKeys,
-}: {
-    base: Record<string, unknown> | undefined
-    resetBase: Record<string, unknown> | undefined
-    resetKeys: string[]
-}) => {
-    const next = {...(base ?? {})}
-    const keysToReset = ["model", ...resetKeys]
-    keysToReset.forEach((key) => {
-        if (resetBase && Object.prototype.hasOwnProperty.call(resetBase, key)) {
-            next[key] = resetBase[key]
-        } else {
-            delete next[key]
-        }
-    })
-    return next
-}
-
-// ============================================================================
-// AGENTA_METADATA HELPERS
-// ============================================================================
-
-/**
- * Recursively strip `agenta_metadata` from tool objects in the parameters tree.
- * Returns a new object safe for display in JSON/YAML view and a map of stripped
- * metadata keyed by stable path so it can be re-attached after editing.
- */
-type MetadataMap = Map<string, unknown>
-
-function stripAgentaMetadata(params: Record<string, unknown>): Record<string, unknown> {
-    return stripRecursive(params) as Record<string, unknown>
-}
-
-function stripRecursive(value: unknown): unknown {
-    if (Array.isArray(value)) {
-        return value.map((item) => stripRecursive(item))
-    }
-    if (value && typeof value === "object") {
-        const obj = value as Record<string, unknown>
-        const result: Record<string, unknown> = {}
-        for (const [k, v] of Object.entries(obj)) {
-            if (k === "agenta_metadata") continue
-            result[k] = stripRecursive(v)
-        }
-        return result
-    }
-    return value
-}
-
-/**
- * Collect all agenta_metadata values from the original parameters,
- * keyed by their JSON path (e.g. "prompt.llm_config.tools.0").
- */
-function collectAgentaMetadata(
-    value: unknown,
-    path = "",
-    map: MetadataMap = new Map(),
-): MetadataMap {
-    if (Array.isArray(value)) {
-        value.forEach((item, i) => collectAgentaMetadata(item, path ? `${path}.${i}` : `${i}`, map))
-    } else if (value && typeof value === "object") {
-        const obj = value as Record<string, unknown>
-        if ("agenta_metadata" in obj) {
-            map.set(path, obj.agenta_metadata)
-        }
-        for (const [k, v] of Object.entries(obj)) {
-            if (k === "agenta_metadata") continue
-            collectAgentaMetadata(v, path ? `${path}.${k}` : k, map)
-        }
-    }
-    return map
-}
-
-/**
- * Re-inject agenta_metadata values into a parsed parameters object
- * using the metadata map collected from the original.
- */
-function reattachAgentaMetadata(value: unknown, metadataMap: MetadataMap, path = ""): unknown {
-    if (Array.isArray(value)) {
-        return value.map((item, i) =>
-            reattachAgentaMetadata(item, metadataMap, path ? `${path}.${i}` : `${i}`),
-        )
-    }
-    if (value && typeof value === "object") {
-        const obj = value as Record<string, unknown>
-        const result: Record<string, unknown> = {}
-        for (const [k, v] of Object.entries(obj)) {
-            result[k] = reattachAgentaMetadata(v, metadataMap, path ? `${path}.${k}` : k)
-        }
-        const meta = metadataMap.get(path)
-        if (meta !== undefined) {
-            result.agenta_metadata = meta
-        }
-        return result
-    }
-    return value
-}
-
-// ============================================================================
-// ATOM MEMOIZATION HELPER
-// ============================================================================
-
-/** Memoize atom factories to return the same atom instance for the same key */
-function memoAtom<T>(factory: (id: string) => Atom<T>): (id: string) => Atom<T> {
-    const cache = new Map<string, Atom<T>>()
-    return (id: string) => {
-        let v = cache.get(id)
-        if (!v) {
-            v = factory(id)
-            cache.set(id, v)
-        }
-        return v
-    }
-}
-
-/** Content-stable wrapper: keep the previous value when a recompute yields equal content. */
-function stableAtom<T>(base: Atom<T>, equals: (a: T, b: T) => boolean): Atom<T> {
-    return selectAtom(base, (v) => v, equals)
-}
-
-// ============================================================================
-// DEFAULT ADAPTER (workflowMolecule — direct molecule access)
-// ============================================================================
-
-const RUNTIME_SELECT_OPTIONS = ["python", "typescript", "javascript"].map((value) => ({
-    label: <span className="capitalize">{value}</span>,
-    value,
-}))
-
-// Virtual section keys for sibling data fields; `["hook","url"]` maps to `data.url`.
-const SIBLING_GROUPS = {
-    hook: ["url", "headers"],
-    code: ["script", "runtime"],
-    schemas: ["parameters", "inputs", "outputs"],
-} as const
-type SiblingGroupKey = keyof typeof SIBLING_GROUPS
-const SIBLING_GROUP_KEYS = Object.keys(SIBLING_GROUPS) as SiblingGroupKey[]
-
-function isSiblingGroupKey(key: unknown): key is SiblingGroupKey {
-    return typeof key === "string" && key in SIBLING_GROUPS
-}
-
-// custom:hook → hook group, custom:code → code group (uri = provider:kind:key:version).
-function allowedSiblingGroup(uri: unknown): SiblingGroupKey | null {
-    if (typeof uri !== "string") return null
-    const [, kind, key] = uri.split(":")
-    if (kind !== "custom") return null
-    return isSiblingGroupKey(key) ? key : null
-}
-
-// Editable schemas (Parameters/Inputs/Outputs) belong only to custom workflows; legacy
-// builtin evaluators carry a fixed, server-owned schema that must not be edited here.
-function isCustomWorkflowUri(uri: unknown): boolean {
-    if (typeof uri !== "string") return false
-    return uri.split(":")[1] === "custom"
-}
-
-// Adapter data: `parameters` + one sibling group (hook/code) as a single section.
-function mergeSiblingFields(
-    config: Record<string, unknown> | null,
-    full: {data?: Record<string, unknown> | null} | null,
-): Record<string, unknown> | null {
-    const fullData = (full?.data ?? null) as Record<string, unknown> | null
-    const group = allowedSiblingGroup(fullData?.uri)
-    const merged: Record<string, unknown> = {parameters: (config ?? {}) as Record<string, unknown>}
-    if (group && fullData) {
-        const SIBLING_FIELD_DEFAULTS: Record<string, unknown> = {headers: {}, runtime: "python"}
-        const fields: Record<string, unknown> = {}
-        for (const field of SIBLING_GROUPS[group]) {
-            fields[field] = fullData[field] ?? SIBLING_FIELD_DEFAULTS[field] ?? ""
-        }
-        merged[group] = fields
-    }
-    // `schemas` nests under data.schemas; editable only for custom workflows.
-    const schemas = fullData?.schemas as Record<string, unknown> | null | undefined
-    if (schemas && typeof schemas === "object" && isCustomWorkflowUri(fullData?.uri)) {
-        const fields: Record<string, unknown> = {}
-        for (const field of SIBLING_GROUPS.schemas) {
-            fields[field] = schemas[field] ?? {}
-        }
-        merged.schemas = fields
-    }
-    if (!config && !group && !merged.schemas) return null
-    return merged
-}
-
-// A path targets a sibling group only when path[0] is a group key AND that group is
-// actually present on the data object. Guards against a legacy *parameter* named
-// "code"/"hook" (path ["code"]) being misrouted as a sibling write.
-function pathTargetsSibling(path: DataPath, data: unknown): boolean {
-    if (path.length === 0 || !isSiblingGroupKey(path[0])) return false
-    const d = data as Record<string, unknown> | null
-    return !!d && path[0] in d
-}
-
-// Renderable if there are parameters OR any sibling group (hook/code/schemas),
-// so sibling-only workflows aren't hidden as "No configuration needed".
-function hasRenderableConfigSections(data: unknown): boolean {
-    if (!data || typeof data !== "object") return false
-    const record = data as Record<string, unknown>
-    if (hasParameters(record)) return true
-    return SIBLING_GROUP_KEYS.some((group) => record[group] !== undefined)
-}
-
-// Best-available data for render/loading decisions: draft-merged data, else server data.
-function pickActiveData<T extends {parameters?: Record<string, unknown>}>(
-    useServerData: boolean,
-    data: T | null,
-    serverData: T | null,
-): T | null {
-    if (useServerData) return serverData
-    if (hasParameters(data)) return data
-    if (hasParameters(serverData)) return serverData
-    return data ?? serverData
-}
-
-// Route tagged `__siblingData` payloads to `actions.update`; otherwise treat as parameter updates.
-const configUpdateRouterAtom = atom(
-    null,
-    (_get, set, id: string, changes: Record<string, unknown>) => {
-        const siblingData = (changes as {__siblingData?: Record<string, unknown>}).__siblingData
-        if (siblingData) {
-            set(workflowMolecule.actions.update, id, {data: siblingData} as Partial<Workflow>)
-            return
-        }
-        set(workflowMolecule.actions.updateConfiguration, id, changes)
-    },
-)
-
-/**
- * Build adapter backed by workflowMolecule.
- *
- * Data mapping:
- * - workflowMolecule.selectors.configuration(id) → `parameters` fields (UI display)
- * - sibling `data.*` fields (script/runtime/url/headers) surfaced alongside,
- *   read from the full resolved data and written via the raw draft action
- * - workflowMolecule.actions.updateConfiguration → parameter writes
- * - workflowMolecule.selectors.parametersSchema(id) → agConfigSchema
- */
-function buildWorkflowMoleculeAdapter(): ConfigSectionMoleculeAdapter {
-    return {
-        atoms: {
-            // deepEqual-stable: upstream selectors mint fresh objects per query flip during boot.
-            data: memoAtom((id: string) =>
-                stableAtom(
-                    atom((get) => {
-                        const config = get(workflowMolecule.selectors.configuration(id))
-                        const full = get(workflowMolecule.selectors.resolvedData(id))
-                        return mergeSiblingFields(config as Record<string, unknown> | null, full)
-                    }),
-                    deepEqual,
-                ),
-            ),
-            serverData: memoAtom((id: string) =>
-                stableAtom(
-                    atom((get) => {
-                        const config = get(workflowMolecule.selectors.serverConfiguration(id))
-                        const full = get(workflowMolecule.selectors.serverData(id))
-                        return mergeSiblingFields(config as Record<string, unknown> | null, full)
-                    }),
-                    deepEqual,
-                ),
-            ),
-            draft: (id: string) => workflowMolecule.atoms.draft(id),
-            isDirty: (id: string) => workflowMolecule.selectors.isDirty(id),
-            schemaQuery: memoAtom((id: string) =>
-                stableAtom(
-                    atom((get) => {
-                        const q = get(workflowMolecule.selectors.query(id))
-                        const rawSchema = get(workflowMolecule.selectors.parametersSchema(id))
-                        const schema = isEntitySchema(rawSchema) ? rawSchema : null
-                        return {
-                            isPending: q.isPending,
-                            isError: q.isError,
-                            error: q.error as Error | null,
-                            data: {agConfigSchema: schema},
-                        }
-                    }),
-                    (a, b) =>
-                        a.isPending === b.isPending &&
-                        a.isError === b.isError &&
-                        a.error === b.error &&
-                        a.data?.agConfigSchema === b.data?.agConfigSchema,
-                ),
-            ),
-            agConfigSchema: memoAtom((id: string) =>
-                atom((get) => {
-                    const schema = get(workflowMolecule.selectors.parametersSchema(id))
-                    return isEntitySchema(schema) ? schema : null
-                }),
-            ),
-        },
-        reducers: {
-            update: configUpdateRouterAtom as WritableAtom<
-                unknown,
-                [id: string, changes: Record<string, unknown>],
-                void
-            >,
-            discard: workflowMolecule.actions.discard,
-        },
-        drillIn: {
-            // Flatten parameter keys + sibling fields to one level (each its own section).
-            getRootData: (data: unknown) => {
-                const d = data as {parameters?: Record<string, unknown>} | null
-                if (!d) return {}
-                const {parameters, ...siblings} = d
-                return {...(parameters ?? {}), ...siblings}
-            },
-            getRootItems: (data: unknown) => {
-                const d = data as {parameters?: Record<string, unknown>} | null
-                const items: {key: string; name: string; value: unknown}[] = []
-                const params = d?.parameters
-                if (params && typeof params === "object") {
-                    for (const [key, value] of Object.entries(params)) {
-                        items.push({key, name: key, value})
-                    }
-                }
-                for (const group of SIBLING_GROUP_KEYS) {
-                    if (d && group in d) {
-                        items.push({key: group, name: group, value: d[group as keyof typeof d]})
-                    }
-                }
-                return items
-            },
-            getValueAtPath: (data: unknown, path: DataPath) => {
-                const d = data as {parameters?: Record<string, unknown>} | null
-                if (!d) return undefined
-                if (pathTargetsSibling(path, d)) return getValueAtPath(d, path)
-                if (!d.parameters) return undefined
-                return getValueAtPath(d.parameters, path)
-            },
-            getChangesFromPath: (data: unknown, path: DataPath, value: unknown) => {
-                const d = data as {parameters?: Record<string, unknown>} | null
-                // Sibling edits emit a tagged __siblingData payload. setValueAtPath
-                // is immutable, so use its return value.
-                if (pathTargetsSibling(path, d)) {
-                    const group = path[0] as SiblingGroupKey
-                    const base = (d as Record<string, unknown>)?.[group] ?? {}
-                    const fields = setValueAtPath(base, path.slice(1), value) as Record<
-                        string,
-                        unknown
-                    >
-                    if (group === "schemas") {
-                        return {__siblingData: {schemas: fields}} as Record<string, unknown>
-                    }
-                    return {__siblingData: fields} as Record<string, unknown>
-                }
-                return setValueAtPath(d?.parameters ?? {}, path, value) as Record<string, unknown>
-            },
-            // rootData flattened: a sibling group edit emits a tagged payload of the
-            // group's fields (paths drop the virtual group key); params emit param keys.
-            getChangesFromRoot: (_entity: unknown, rootData: unknown, path: DataPath) => {
-                const root = {...(rootData as Record<string, unknown>)}
-                if (pathTargetsSibling(path, _entity)) {
-                    const group = path[0] as SiblingGroupKey
-                    const fields = (root[group] ?? {}) as Record<string, unknown>
-                    // schemas nests under data.schemas; hook/code fields sit flat on data.
-                    if (group === "schemas") {
-                        return {__siblingData: {schemas: {...fields}}} as Record<string, unknown>
-                    }
-                    return {__siblingData: {...fields}} as Record<string, unknown>
-                }
-                // Only strip keys that are real sibling groups on the entity, so a
-                // legacy parameter named "code"/"hook" survives in the param payload.
-                const entity = _entity as Record<string, unknown> | null
-                for (const group of SIBLING_GROUP_KEYS) {
-                    if (entity && group in entity) delete root[group]
-                }
-                return root
-            },
-        },
-        selectors: {
-            schemaAtPath: memoAtom((key: string) => {
-                // key is serialized "{id}:{path}" — parse it
-                // But the interface takes {id, path}, so we use a wrapper below
-                return atom(() => null)
-            }) as unknown as ConfigSectionMoleculeAdapter["selectors"]["schemaAtPath"],
-        },
-    }
-}
-
-// Minimal object schema so the section is recognized; the hook/code body renders
-// via HookCodeConfigControl, not the schema renderer, so no field schemas needed.
-function siblingSchemaAtPath(path: (string | number)[]): PathSchema | null {
-    const group = path[0]
-    if (!isSiblingGroupKey(group)) return null
-    if (path.length === 1) {
-        return {type: "object", title: group} as EntitySchemaProperty
-    }
-    return null
-}
-
-/** Wrap schemaAtPath to work with the adapter's (id, path) → atom interface */
-const moleculeSchemaAtPathCache = new Map<string, Atom<unknown>>()
-function moleculeSchemaAtPath(params: {id: string; path: (string | number)[]}): Atom<unknown> {
-    const key = `${params.id}:${params.path.join(".")}`
-    let cached = moleculeSchemaAtPathCache.get(key)
-    if (!cached) {
-        cached = atom((get) => {
-            const full = get(workflowMolecule.selectors.resolvedData(params.id))
-            const adapterData = mergeSiblingFields(null, full)
-            if (pathTargetsSibling(params.path as DataPath, adapterData)) {
-                return siblingSchemaAtPath(params.path)
-            }
-            const schema = get(workflowMolecule.selectors.parametersSchema(params.id))
-            if (!isEntitySchema(schema)) return null
-            const resolved = getSchemaAtPathUtil(schema, params.path) ?? null
-            if (params.path.length === 0) {
-                console.debug("[PlaygroundConfigSection] root schema", params.id.slice(0, 8), {
-                    schemaType: schema?.type,
-                    schemaHasProperties: !!schema?.properties,
-                    schemaPropertyKeys: schema?.properties
-                        ? Object.keys(schema.properties as Record<string, unknown>)
-                        : null,
-                    resolvedType: resolved?.type,
-                    resolvedPropertyKeys: resolved?.properties
-                        ? Object.keys(resolved.properties as Record<string, unknown>)
-                        : null,
-                })
-            }
-            return resolved
-        })
-        moleculeSchemaAtPathCache.set(key, cached)
-    }
-    return cached
-}
-
-function buildDefaultAdapter(): ConfigSectionMoleculeAdapter {
-    const base = buildWorkflowMoleculeAdapter()
-    return {
-        ...base,
-        selectors: {
-            schemaAtPath: moleculeSchemaAtPath,
-        },
-    }
-}
-
-const defaultAdapter = buildDefaultAdapter()
-
-// ============================================================================
-// COMPONENT
-// ============================================================================
-
-/** Evaluator preset definition */
-export interface EvaluatorPresetConfig {
-    key: string
-    name: string
-    values: Record<string, unknown>
-}
-
-export type ConfigViewMode = "form" | "json" | "yaml"
-
-export interface PlaygroundConfigSectionProps {
-    revisionId: string
-    disabled?: boolean
-    useServerData?: boolean
-    className?: string
-    /** Optional molecule adapter — defaults to workflowMolecule */
-    moleculeAdapter?: ConfigSectionMoleculeAdapter
-    /** Called when the user clicks "Refine prompt with AI" on a prompt section header */
-    onRefinePrompt?: (promptKey: string) => void
-    /** View mode controlled from parent (form/json/yaml) */
-    viewMode?: ConfigViewMode
-    /**
-     * Top offset (px) for the sticky section headers. Defaults to 48 to clear
-     * the sticky `PlaygroundVariantConfigHeader` (h-[48px], `sticky top-0`) that
-     * sits above the config in the full playground. In the embedded drawer that
-     * header is rendered non-sticky (`grow`), so there is nothing to clear —
-     * pass 0 there to keep the section headers flush with the scroll top instead
-     * of floating 48px down into the editor content.
-     */
-    stickyHeaderTop?: number
-    /**
-     * Rendered instead of the generic pulse boxes while the config/schema is
-     * loading. Lets the caller show a layout-matched skeleton (e.g. the agent
-     * section-row list) when it knows the entity's shape before the data lands.
-     */
-    loadingFallback?: React.ReactNode
-}
+export type {
+    ConfigSectionMoleculeAdapter,
+    ConfigViewMode,
+    EvaluatorPresetConfig,
+    PlaygroundConfigSectionProps,
+} from "./PlaygroundConfigSection/types"
 
 function PlaygroundConfigSection({
     revisionId: targetRevisionId,
@@ -926,969 +324,46 @@ function PlaygroundConfigSection({
     }, [])
 
     // ========== COMBINED MODEL / FALLBACK / RETRY CONFIG POPOVER ==========
-    const [isModelConfigOpen, setIsModelConfigOpen] = useState(false)
-    const [activeConfigureTab, setActiveConfigureTab] = useState<ConfigureTabKey>("model")
-    const [fallbackDetail, setFallbackDetail] = useState<FallbackDetailState | null>(null)
-
-    // Extract model + LLM config info from prompt section.
-    //
-    // Supports two schema shapes:
-    // - Legacy: parameters.prompt.{messages, llm_config.{model, temperature, ...}}
-    // - Canonical (llm catalog): parameters.{messages, llms[{model, temperature, ...}]}
-    //
-    // For the canonical shape, the root parameters object IS the prompt equivalent.
-    const promptModelInfo = useMemo(() => {
-        const hasNestedPrompt = !!parameters.prompt
-        const hasRootMessages = Array.isArray(parameters.messages)
-
-        const promptValue = hasNestedPrompt
-            ? (parameters.prompt as Record<string, unknown>)
-            : hasRootMessages
-              ? parameters
-              : null
-        if (!promptValue) return null
-
-        const promptSchema = schema?.properties
-            ? hasNestedPrompt
-                ? ((schema.properties as Record<string, EntitySchemaProperty>).prompt ?? null)
-                : hasRootMessages
-                  ? schema
-                  : null
-            : null
-        const modelSchema = getModelSchema(promptSchema as EntitySchemaProperty | null)
-        const optionsResult = getOptionsFromSchema(modelSchema)
-        const modelOptions = optionsResult?.options ?? []
-
-        const llmConfigValue = getLLMConfigValue(promptValue)
-        const currentModel = llmConfigValue?.model as string | undefined
-
-        // Extract LLM config property schemas for sliders
-        const llmConfigProps = getLLMConfigProperties(promptSchema as EntitySchemaProperty | null)
-        const promptSchemaProps = ((promptSchema as EntitySchemaProperty | null)?.properties ??
-            {}) as Record<string, EntitySchemaProperty>
-
-        return {
-            modelSchema,
-            modelOptions,
-            currentModel,
-            promptValue,
-            promptSchemaProps,
-            llmConfigValue,
-            llmConfigProps,
-            isRootLevel: !hasNestedPrompt && hasRootMessages,
-        }
-    }, [parameters, schema])
-
-    // Helper to update a key inside the LLM config.
-    // Supports three structures:
-    // - Legacy nested: parameters.prompt.llm_config.{key}
-    // - Legacy flat: parameters.prompt.{key}
-    // - Canonical: parameters.llms[0].{key}
-    const updatePromptLLMConfigKey = useCallback(
-        (key: string, newValue: unknown) => {
-            if (disabled || !activeData) return
-
-            // Canonical: llms array at root level
-            if (Array.isArray(parameters.llms)) {
-                const currentLlms = parameters.llms as Record<string, unknown>[]
-                const updatedFirst = updateConfigKey(currentLlms[0], key, newValue)
-                dispatchUpdate(revisionId, {
-                    ...parameters,
-                    llms: [updatedFirst, ...currentLlms.slice(1)],
-                })
-                return
-            }
-
-            // Canonical/root-level prompt without an llms array.
-            if (promptModelInfo?.isRootLevel) {
-                const nextParameters = updateConfigKey(parameters, key, newValue)
-                dispatchUpdate(revisionId, nextParameters)
-                return
-            }
-
-            // Legacy: prompt.llm_config or prompt.{key}
-            const currentPrompt = (parameters.prompt as Record<string, unknown>) || {}
-            const hasNestedLLMConfig = currentPrompt.llm_config || currentPrompt.llmConfig
-
-            let updatedPrompt
-            if (hasNestedLLMConfig) {
-                const llmConfigKey = currentPrompt.llm_config ? "llm_config" : "llmConfig"
-                updatedPrompt = {
-                    ...currentPrompt,
-                    [llmConfigKey]: updateConfigKey(
-                        currentPrompt[llmConfigKey] as Record<string, unknown> | undefined,
-                        key,
-                        newValue,
-                    ),
-                }
-            } else {
-                updatedPrompt = updateConfigKey(currentPrompt, key, newValue)
-            }
-
-            dispatchUpdate(revisionId, {
-                ...parameters,
-                prompt: updatedPrompt,
-            })
-        },
-        [disabled, activeData, parameters, revisionId, dispatchUpdate, promptModelInfo],
-    )
-
-    const updatePromptRootFields = useCallback(
-        (changes: Record<string, unknown>) => {
-            if (disabled || !activeData || !promptModelInfo) return
-
-            const applyChanges = (base: Record<string, unknown>) => {
-                const next = {...base}
-                for (const [key, value] of Object.entries(changes)) {
-                    if (value === null || value === undefined) {
-                        delete next[key]
-                    } else {
-                        next[key] = value
-                    }
-                }
-                return next
-            }
-
-            if (promptModelInfo.isRootLevel) {
-                dispatchUpdate(revisionId, applyChanges(parameters))
-                return
-            }
-
-            const currentPrompt = (parameters.prompt as Record<string, unknown>) || {}
-            dispatchUpdate(revisionId, {
-                ...parameters,
-                prompt: applyChanges(currentPrompt),
-            })
-        },
-        [disabled, activeData, promptModelInfo, dispatchUpdate, revisionId, parameters],
-    )
-
-    const updatePromptRootField = useCallback(
-        (key: string, nextValue: unknown) => {
-            updatePromptRootFields({[key]: nextValue})
-        },
-        [updatePromptRootFields],
-    )
-
-    const fallbackConfigs = useMemo(() => {
-        const raw = promptModelInfo?.promptValue.fallback_configs
-        return Array.isArray(raw) ? (raw as Record<string, unknown>[]) : []
-    }, [promptModelInfo])
-    const [fallbackConfigKeys, setFallbackConfigKeys] = useState<string[]>([])
-    useEffect(() => {
-        setFallbackConfigKeys((currentKeys) => {
-            if (currentKeys.length === fallbackConfigs.length) return currentKeys
-            if (currentKeys.length > fallbackConfigs.length) {
-                return currentKeys.slice(0, fallbackConfigs.length)
-            }
-
-            return [
-                ...currentKeys,
-                ...Array.from(
-                    {length: fallbackConfigs.length - currentKeys.length},
-                    createFallbackConfigKey,
-                ),
-            ]
-        })
-    }, [fallbackConfigs.length])
-
-    const retryConfig = useMemo(() => {
-        const raw = promptModelInfo?.promptValue.retry_config
-        return raw && typeof raw === "object" && !Array.isArray(raw)
-            ? (raw as Record<string, unknown>)
-            : {}
-    }, [promptModelInfo])
-
-    const effectiveRetryConfig = useMemo(
-        () => ({
-            max_retries:
-                typeof retryConfig.max_retries === "number"
-                    ? retryConfig.max_retries
-                    : DEFAULT_RETRY_CONFIG.max_retries,
-            base_delay:
-                typeof retryConfig.base_delay === "number"
-                    ? retryConfig.base_delay
-                    : DEFAULT_RETRY_CONFIG.base_delay,
-        }),
-        [retryConfig],
-    )
-
-    const hasPromptExtensionFields = useMemo(() => {
-        if (!promptModelInfo) return false
-        return PROMPT_EXTENSION_KEYS.some(
-            (key) => key in promptModelInfo.promptSchemaProps || key in promptModelInfo.promptValue,
-        )
-    }, [promptModelInfo])
-
-    useEffect(() => {
-        if (!hasPromptExtensionFields && activeConfigureTab !== "model") {
-            setActiveConfigureTab("model")
-            setFallbackDetail(null)
-        }
-    }, [activeConfigureTab, hasPromptExtensionFields])
-
-    const fallbackPolicyOptions = useMemo(() => {
-        const schema = promptModelInfo?.promptSchemaProps.fallback_policy as
-            | {enum?: unknown[]; "x-ag-metadata"?: Record<string, {description?: string}>}
-            | undefined
-        const metadata = schema?.["x-ag-metadata"] ?? {}
-        const enumValues = schema?.enum
-        const values =
-            Array.isArray(enumValues) && enumValues.length > 0
-                ? enumValues.map((v) => String(v))
-                : FALLBACK_POLICY_OPTIONS.map((o) => o.value)
-        return values.map((value) => ({
-            label: formatLabel(value),
-            value,
-            description: metadata[value]?.description,
-        }))
-    }, [promptModelInfo])
-
-    const retryPolicyOptions = useMemo(() => {
-        const schema = promptModelInfo?.promptSchemaProps.retry_policy as
-            | {enum?: unknown[]; "x-ag-metadata"?: Record<string, {description?: string}>}
-            | undefined
-        const metadata = schema?.["x-ag-metadata"] ?? {}
-        const enumValues = schema?.enum
-        const values =
-            Array.isArray(enumValues) && enumValues.length > 0
-                ? enumValues.map((v) => String(v))
-                : RETRY_POLICY_OPTIONS.map((o) => o.value)
-        return values.map((value) => ({
-            label: formatLabel(value),
-            value,
-            description: metadata[value]?.description,
-        }))
-    }, [promptModelInfo])
-
-    const fallbackModelOptions = useMemo(
-        () => [
-            ...(llmProviderConfig?.extraOptionGroups ?? []),
-            ...(promptModelInfo?.modelOptions ?? []),
-        ],
-        [llmProviderConfig?.extraOptionGroups, promptModelInfo?.modelOptions],
-    )
-
-    const handleRetryConfigFieldChange = useCallback(
-        (key: "max_retries" | "base_delay", nextValue: number | null) => {
-            const nextMaxRetries =
-                key === "max_retries"
-                    ? nextValue
-                    : typeof retryConfig.max_retries === "number"
-                      ? retryConfig.max_retries
-                      : DEFAULT_RETRY_CONFIG.max_retries
-
-            if (typeof nextMaxRetries !== "number" || nextMaxRetries <= 0) {
-                updatePromptRootFields({
-                    retry_config: null,
-                    retry_policy: null,
-                })
-                return
-            }
-
-            const nextBaseDelay =
-                key === "base_delay"
-                    ? nextValue
-                    : typeof retryConfig.base_delay === "number"
-                      ? retryConfig.base_delay
-                      : DEFAULT_RETRY_CONFIG.base_delay
-            const nextRetryConfig: Record<string, unknown> = {
-                max_retries: nextMaxRetries,
-            }
-            if (typeof nextBaseDelay === "number") {
-                nextRetryConfig.base_delay = nextBaseDelay
-            }
-
-            updatePromptRootField("retry_config", nextRetryConfig)
-        },
-        [retryConfig, updatePromptRootField, updatePromptRootFields],
-    )
-
-    const handleAddFallbackModel = useCallback(() => {
-        const primaryModel =
-            typeof promptModelInfo?.llmConfigValue?.model === "string"
-                ? promptModelInfo.llmConfigValue.model
-                : ""
-        setFallbackDetail({
-            mode: "new",
-            index: null,
-            draft: {model: primaryModel || "gpt-4o-mini"},
-        })
-    }, [promptModelInfo])
-
-    const handleEditFallbackModel = useCallback(
-        (index: number) => {
-            setFallbackDetail({
-                mode: "edit",
-                index,
-                draft: {...(fallbackConfigs[index] ?? {})},
-            })
-        },
-        [fallbackConfigs],
-    )
-
-    const handleFallbackDetailChange = useCallback((key: string, nextValue: unknown) => {
-        setFallbackDetail((current) => {
-            if (!current) return current
-            const nextDraft = {...current.draft}
-            if (nextValue === null || nextValue === undefined) {
-                delete nextDraft[key]
-            } else {
-                nextDraft[key] = nextValue
-            }
-            return {...current, draft: nextDraft}
-        })
-    }, [])
-
-    const handleCommitFallbackDetail = useCallback(() => {
-        if (!fallbackDetail) return
-        const nextConfigs =
-            fallbackDetail.mode === "edit" && fallbackDetail.index !== null
-                ? fallbackConfigs.map((config, configIndex) =>
-                      configIndex === fallbackDetail.index ? fallbackDetail.draft : config,
-                  )
-                : [...fallbackConfigs, fallbackDetail.draft]
-
-        updatePromptRootField("fallback_configs", nextConfigs.length > 0 ? nextConfigs : null)
-        setFallbackDetail(null)
-    }, [fallbackConfigs, fallbackDetail, updatePromptRootField])
-
-    const handleRemoveFallbackModel = useCallback(
-        (index: number) => {
-            const nextConfigs = fallbackConfigs.filter((_, configIndex) => configIndex !== index)
-            setFallbackConfigKeys((currentKeys) =>
-                currentKeys.filter((_, configIndex) => configIndex !== index),
-            )
-            if (nextConfigs.length === 0) {
-                updatePromptRootFields({
-                    fallback_configs: null,
-                    fallback_policy: null,
-                })
-                return
-            }
-
-            updatePromptRootField("fallback_configs", nextConfigs)
-        },
-        [fallbackConfigs, updatePromptRootField, updatePromptRootFields],
-    )
-
-    const handleResetFallbackPolicy = useCallback(() => {
-        updatePromptRootFields({
-            fallback_policy: null,
-            fallback_configs: null,
-        })
-        setFallbackConfigKeys([])
-        setFallbackDetail(null)
-    }, [updatePromptRootFields])
-
-    const handleResetRetryPolicy = useCallback(() => {
-        updatePromptRootFields({
-            retry_config: null,
-            retry_policy: null,
-        })
-    }, [updatePromptRootFields])
-
-    const handleResetPrimaryModelConfig = useCallback(() => {
-        if (disabled || !activeData || !promptModelInfo) return
-
-        const serverParameters = (serverData?.parameters ?? {}) as Record<string, unknown>
-        const getServerLLMConfig = () => {
-            if (Array.isArray(parameters.llms)) {
-                const serverLlms = serverParameters.llms as Record<string, unknown>[] | undefined
-                return serverLlms?.[0]
-            }
-
-            if (promptModelInfo.isRootLevel) {
-                return serverParameters
-            }
-
-            const serverPrompt = serverParameters.prompt as Record<string, unknown> | undefined
-            return getLLMConfigValue(serverPrompt)
-        }
-
-        const resetBase = getServerLLMConfig()
-        const resetKeys = getResettableLLMConfigKeys(promptModelInfo.llmConfigProps)
-        const resetLLMConfig = (base: Record<string, unknown> | undefined) => {
-            return resetLLMParameterFields({base, resetBase, resetKeys})
-        }
-
-        if (Array.isArray(parameters.llms)) {
-            const currentLlms = parameters.llms as Record<string, unknown>[]
-            dispatchUpdate(revisionId, {
-                ...parameters,
-                llms: [resetLLMConfig(currentLlms[0]), ...currentLlms.slice(1)],
-            })
-            return
-        }
-
-        if (promptModelInfo.isRootLevel) {
-            dispatchUpdate(revisionId, resetLLMConfig(parameters))
-            return
-        }
-
-        const currentPrompt = (parameters.prompt as Record<string, unknown>) || {}
-        const hasNestedLLMConfig = currentPrompt.llm_config || currentPrompt.llmConfig
-
-        if (hasNestedLLMConfig) {
-            const llmKey = currentPrompt.llm_config ? "llm_config" : "llmConfig"
-            dispatchUpdate(revisionId, {
-                ...parameters,
-                prompt: {
-                    ...currentPrompt,
-                    [llmKey]: resetLLMConfig(
-                        currentPrompt[llmKey] as Record<string, unknown> | undefined,
-                    ),
-                },
-            })
-            return
-        }
-
-        dispatchUpdate(revisionId, {
-            ...parameters,
-            prompt: resetLLMConfig(currentPrompt),
-        })
-    }, [activeData, disabled, dispatchUpdate, parameters, promptModelInfo, revisionId, serverData])
-
-    const handleActiveConfigureReset = useCallback(() => {
-        if (fallbackDetail || activeConfigureTab === "fallback") {
-            handleResetFallbackPolicy()
-            return
-        }
-        if (activeConfigureTab === "retry") {
-            handleResetRetryPolicy()
-            return
-        }
-        handleResetPrimaryModelConfig()
-    }, [
-        activeConfigureTab,
-        fallbackDetail,
-        handleResetFallbackPolicy,
-        handleResetPrimaryModelConfig,
-        handleResetRetryPolicy,
-    ])
-
-    const handleConfigureOpenChange = useCallback(
-        (open: boolean) => {
-            setIsModelConfigOpen(open)
-            if (!open && fallbackDetail) {
-                handleCommitFallbackDetail()
-            }
-        },
-        [fallbackDetail, handleCommitFallbackDetail],
-    )
-
-    const handleConfigureTabChange = useCallback((key: string) => {
-        setActiveConfigureTab(key as ConfigureTabKey)
-        setFallbackDetail(null)
-    }, [])
-
-    const handlePrimaryModelConfigChange = useCallback(
-        (key: string, next: unknown) => updatePromptLLMConfigKey(key, next),
-        [updatePromptLLMConfigKey],
-    )
-
-    const handleFallbackPolicyChange = useCallback(
-        (nextValue: string | null) => updatePromptRootField("fallback_policy", nextValue),
-        [updatePromptRootField],
-    )
-
-    const handleRetryPolicyChange = useCallback(
-        (nextValue: string | null) => updatePromptRootField("retry_policy", nextValue),
-        [updatePromptRootField],
-    )
-
-    const configurePopoverContent = useMemo(
-        () => (
-            <div className="w-[320px] max-h-[550px] overflow-hidden rounded bg-[var(--ag-c-FFFFFF)]">
-                <div className="flex items-center justify-between gap-3 border-0 border-b border-solid border-[var(--ag-rgba-051729-08)] bg-[var(--ag-c-F6F8FA)] px-3 py-2">
-                    <div className="flex items-center gap-2 min-w-0">
-                        {fallbackDetail && (
-                            <Button
-                                size="small"
-                                type="text"
-                                icon={<ArrowLeft size={16} />}
-                                onClick={handleCommitFallbackDetail}
-                                disabled={disabled}
-                                aria-label="Back to fallback models"
-                                className="flex items-center justify-center"
-                            />
-                        )}
-                        <Typography.Text className="truncate font-medium">
-                            {fallbackDetail
-                                ? fallbackDetail.mode === "new"
-                                    ? "Add Fallback Model"
-                                    : "Edit Fallback Model"
-                                : "Configure"}
-                        </Typography.Text>
-                        {fallbackDetail?.mode === "new" && (
-                            <span className="rounded bg-blue-50 px-2 py-0.5 text-xs text-blue-600">
-                                new
-                            </span>
-                        )}
-                    </div>
-                    <Button size="small" onClick={handleActiveConfigureReset} disabled={disabled}>
-                        Reset to default
-                    </Button>
-                </div>
-
-                {fallbackDetail ? (
-                    <div className="max-h-[498px] overflow-y-auto px-3 py-3">
-                        <ModelConfigEditor
-                            value={fallbackDetail.draft}
-                            onChange={handleFallbackDetailChange}
-                            llmConfigProps={promptModelInfo?.llmConfigProps ?? {}}
-                            modelOptions={fallbackModelOptions}
-                            footerContent={llmProviderConfig?.footerContent}
-                            disabled={disabled}
-                            excludeKeys={PROMPT_EXTENSION_KEYS}
-                        />
-                    </div>
-                ) : (
-                    <Tabs
-                        activeKey={activeConfigureTab}
-                        onChange={handleConfigureTabChange}
-                        className="[&_.ant-tabs-nav]:!mb-0 [&_.ant-tabs-nav]:!bg-[var(--ag-c-F6F8FA)] [&_.ant-tabs-nav]:!px-0 [&_.ant-tabs-nav-wrap]:!w-full [&_.ant-tabs-nav-list]:!w-full [&_.ant-tabs-tab]:!basis-0 [&_.ant-tabs-tab]:!flex-1 [&_.ant-tabs-tab]:!justify-center [&_.ant-tabs-tab]:!mx-0 [&_.ant-tabs-tab-btn]:!mx-auto [&_.ant-tabs-content-holder]:max-h-[452px] [&_.ant-tabs-content-holder]:overflow-y-auto [&_.ant-tabs-content-holder]:px-3 [&_.ant-tabs-content-holder]:py-3"
-                        items={[
-                            {
-                                key: "model",
-                                label: "Model",
-                                children: promptModelInfo ? (
-                                    <ModelConfigEditor
-                                        value={
-                                            (promptModelInfo.llmConfigValue ?? {}) as Record<
-                                                string,
-                                                unknown
-                                            >
-                                        }
-                                        onChange={handlePrimaryModelConfigChange}
-                                        llmConfigProps={promptModelInfo.llmConfigProps}
-                                        modelOptions={[
-                                            ...(llmProviderConfig?.extraOptionGroups ?? []),
-                                            ...promptModelInfo.modelOptions,
-                                        ]}
-                                        footerContent={llmProviderConfig?.footerContent}
-                                        disabled={disabled}
-                                        excludeKeys={PROMPT_EXTENSION_KEYS}
-                                    />
-                                ) : null,
-                            },
-                            ...(hasPromptExtensionFields
-                                ? [
-                                      {
-                                          key: "fallback",
-                                          label: "Fallback",
-                                          children: (
-                                              <FallbackConfigTab
-                                                  fallbackPolicy={
-                                                      (promptModelInfo?.promptValue
-                                                          .fallback_policy as
-                                                          | string
-                                                          | null
-                                                          | undefined) ?? null
-                                                  }
-                                                  fallbackConfigs={fallbackConfigs}
-                                                  fallbackConfigKeys={fallbackConfigKeys}
-                                                  fallbackPolicyOptions={fallbackPolicyOptions}
-                                                  fallbackPolicySchema={
-                                                      promptModelInfo?.promptSchemaProps
-                                                          .fallback_policy as
-                                                          | EntitySchemaProperty
-                                                          | undefined
-                                                  }
-                                                  fallbackConfigsSchema={
-                                                      promptModelInfo?.promptSchemaProps
-                                                          .fallback_configs as
-                                                          | EntitySchemaProperty
-                                                          | undefined
-                                                  }
-                                                  onPolicyChange={handleFallbackPolicyChange}
-                                                  onAddFallbackModel={handleAddFallbackModel}
-                                                  onEditFallbackModel={handleEditFallbackModel}
-                                                  onRemoveFallbackModel={handleRemoveFallbackModel}
-                                                  disabled={disabled}
-                                              />
-                                          ),
-                                      },
-                                      {
-                                          key: "retry",
-                                          label: "Retry",
-                                          children: (
-                                              <RetryConfigTab
-                                                  retryPolicy={
-                                                      (promptModelInfo?.promptValue.retry_policy as
-                                                          | string
-                                                          | null
-                                                          | undefined) ?? null
-                                                  }
-                                                  retryPolicyOptions={retryPolicyOptions}
-                                                  retryPolicySchema={
-                                                      promptModelInfo?.promptSchemaProps
-                                                          .retry_policy as
-                                                          | EntitySchemaProperty
-                                                          | undefined
-                                                  }
-                                                  retryConfigSchema={
-                                                      promptModelInfo?.promptSchemaProps
-                                                          .retry_config as
-                                                          | EntitySchemaProperty
-                                                          | undefined
-                                                  }
-                                                  maxRetries={effectiveRetryConfig.max_retries}
-                                                  baseDelay={effectiveRetryConfig.base_delay}
-                                                  onPolicyChange={handleRetryPolicyChange}
-                                                  onConfigFieldChange={handleRetryConfigFieldChange}
-                                                  disabled={disabled}
-                                              />
-                                          ),
-                                      },
-                                  ]
-                                : []),
-                        ]}
-                    />
-                )}
-            </div>
-        ),
-        [
-            activeConfigureTab,
-            disabled,
-            effectiveRetryConfig.base_delay,
-            effectiveRetryConfig.max_retries,
-            fallbackConfigKeys,
-            fallbackConfigs,
-            fallbackDetail,
-            fallbackModelOptions,
-            fallbackPolicyOptions,
-            handleActiveConfigureReset,
-            handleAddFallbackModel,
-            handleCommitFallbackDetail,
-            handleConfigureTabChange,
-            handleEditFallbackModel,
-            handleFallbackDetailChange,
-            handleFallbackPolicyChange,
-            handlePrimaryModelConfigChange,
-            handleRemoveFallbackModel,
-            handleRetryConfigFieldChange,
-            handleRetryPolicyChange,
-            hasPromptExtensionFields,
-            llmProviderConfig?.extraOptionGroups,
-            llmProviderConfig?.footerContent,
-            promptModelInfo,
-            retryPolicyOptions,
-        ],
-    )
-
-    // ========== FIELD ACTIONS SLOT ==========
-    const fieldActionsSlot = useCallback((props: FieldActionsSlotProps) => {
-        if (props.path.length === 1) return null
-        return props.defaultRender()
-    }, [])
-
-    // ========== FIELD HEADER SLOT ==========
-    const fieldHeaderSlot = useCallback(
-        (props: FieldHeaderSlotProps) => {
-            const fieldKey = String(props.field.key)
-            const isTopLevel = props.path.length === 1
-            if (!isTopLevel) return props.defaultRender()
-
-            // Hide llms header when handled by the model popover
-            if (fieldKey === "llms" && promptModelInfo?.isRootLevel) {
-                return null
-            }
-
-            // Sibling group (hook/code) always gets a section header.
-            const isSibling = isPresentSiblingGroup(fieldKey)
-
-            // Scalar/array params render inline (no header); only objects get one.
-            const fieldValue = isSibling ? siblingGroups[fieldKey] : parameters[fieldKey]
-            if (
-                !isSibling &&
-                (fieldValue === null ||
-                    fieldValue === undefined ||
-                    typeof fieldValue !== "object" ||
-                    Array.isArray(fieldValue))
-            ) {
-                return null
-            }
-
-            // Show model popover on "prompt" section header.
-            // For canonical schemas (no prompt wrapper), show it on the "messages" section.
-            const isPromptWithPopover =
-                (fieldKey === "prompt" ||
-                    (fieldKey === "messages" && promptModelInfo?.isRootLevel)) &&
-                !!promptModelInfo
-            const isCollapsed = !!collapsedSections[fieldKey]
-
-            // Determine if this field has messages for the refine button
-            const hasMessages =
-                !!fieldValue &&
-                typeof fieldValue === "object" &&
-                !Array.isArray(fieldValue) &&
-                Array.isArray((fieldValue as Record<string, unknown>).messages)
-
-            // Get display label from schema title if available, falling back to formatLabel
-            const fieldSchema = schema?.properties
-                ? (schema.properties as Record<string, Record<string, unknown>>)[fieldKey]
-                : null
-            const schemaTitle = fieldSchema?.title as string | undefined
-
-            // Agent workflows render the whole agent block as the panel body (its own sections each
-            // have headers and drawers), so the redundant top-level "Agent" collapse header is
-            // suppressed. Detected via the same schema marker AgentTemplateControl dispatches on —
-            // both the new `agent-template` and the legacy `agent_config` name, so it holds across
-            // the rename / un-migrated data.
-            const isAgentMarker = (v: unknown) => v === "agent-template" || v === "agent_config"
-            if (
-                isAgentMarker(fieldSchema?.["x-ag-type-ref"]) ||
-                isAgentMarker(fieldSchema?.["x-ag-type"])
-            ) {
-                return null
-            }
-
-            const displayLabel = schemaTitle
-                ? schemaTitle.includes(" ")
-                    ? schemaTitle
-                    : formatLabel(schemaTitle)
-                : formatLabel(fieldKey)
-
-            return (
-                <div
-                    className={clsx(
-                        "flex items-center justify-between w-full px-3 py-2 bg-[var(--ag-c-FAFAFB)] cursor-pointer select-none sticky z-[2]",
-                        // Space above the code/hook section when params precede it.
-                        isSibling && hasParameters(activeData) && "mt-4",
-                    )}
-                    style={{top: stickyHeaderTop}}
-                    onClick={() => toggleSection(fieldKey)}
-                >
-                    <div className="flex items-center gap-1">
-                        <span className="text-[var(--ag-rgba-051729-45)] flex items-center">
-                            {isCollapsed ? (
-                                <CaretRight size={14} weight="bold" />
-                            ) : (
-                                <CaretDown size={14} weight="bold" />
-                            )}
-                        </span>
-                        <span className="capitalize font-medium text-sm">{displayLabel}</span>
-                    </div>
-
-                    {isPromptWithPopover && promptModelInfo && (
-                        <div
-                            onClick={(e) => e.stopPropagation()}
-                            className="flex items-center gap-2 flex-shrink-0"
-                        >
-                            {/* Note: the prompt template_format picker is NOT
-                             *  rendered here. Per Mahmoud's QA on the
-                             *  mustache rollout (Slack #release-v100, 2026-
-                             *  05-28), the picker only belongs to the right
-                             *  of the output-type / response-format control
-                             *  inside `PromptSchemaControl`, not in this
-                             *  section header. Adding it here was a
-                             *  duplicate placement (commit 2854349c47) and
-                             *  has been reverted. The original picker in
-                             *  `PromptSchemaControl` continues to handle
-                             *  the format choice for the playground. */}
-                            {!disabled && onRefinePrompt && hasMessages && (
-                                <Tooltip title="Refine prompt with AI">
-                                    <Button
-                                        type="text"
-                                        size="small"
-                                        icon={<MagicWand size={16} aria-hidden="true" />}
-                                        onClick={() => onRefinePrompt(fieldKey)}
-                                        aria-label="Refine prompt with AI"
-                                        className="flex items-center justify-center opacity-60 hover:opacity-100"
-                                    />
-                                </Tooltip>
-                            )}
-                            <Popover
-                                trigger="click"
-                                open={isModelConfigOpen}
-                                onOpenChange={handleConfigureOpenChange}
-                                placement="bottomRight"
-                                arrow={false}
-                                content={configurePopoverContent}
-                                overlayInnerStyle={{padding: 0}}
-                            >
-                                <Button size="small" type="default">
-                                    {promptModelInfo.currentModel || "Select model"}
-                                    <CaretDown size={12} />
-                                </Button>
-                            </Popover>
-                        </div>
-                    )}
-
-                    {/* Code: runtime picker in section header (like the model picker). */}
-                    {isSibling && fieldKey === "code" && (
-                        <div
-                            onClick={(e) => e.stopPropagation()}
-                            className="flex items-center gap-2 flex-shrink-0"
-                        >
-                            <Dropdown
-                                trigger={["click"]}
-                                disabled={disabled}
-                                menu={{
-                                    items: RUNTIME_SELECT_OPTIONS.map((o) => ({
-                                        key: o.value,
-                                        label: o.label,
-                                    })),
-                                    selectedKeys: codeRuntime ? [codeRuntime] : [],
-                                    onClick: ({key}) => handleRuntimeChange(key),
-                                }}
-                            >
-                                <Button size="small" type="default">
-                                    {RUNTIME_SELECT_OPTIONS.find((o) => o.value === codeRuntime)
-                                        ?.label ?? "Select runtime"}
-                                    <CaretDown size={12} />
-                                </Button>
-                            </Dropdown>
-                        </div>
-                    )}
-
-                    {/* Feedback config: Advanced Mode toggle in section header */}
-                    {fieldKey === "feedback_config" && (
-                        <div
-                            onClick={(e) => e.stopPropagation()}
-                            className="flex items-center flex-shrink-0"
-                        >
-                            <Button
-                                size="small"
-                                type="text"
-                                onClick={() =>
-                                    setFeedbackMode(feedbackMode === "basic" ? "advanced" : "basic")
-                                }
-                                disabled={disabled}
-                                className="text-xs text-gray-500"
-                            >
-                                {feedbackMode === "basic" ? "Advanced" : "Basic"}
-                            </Button>
-                        </div>
-                    )}
-                </div>
-            )
-        },
-        [
-            collapsedSections,
-            toggleSection,
-            promptModelInfo,
-            isModelConfigOpen,
-            disabled,
-            parameters,
-            schema,
-            revisionId,
-            feedbackMode,
-            setFeedbackMode,
-            handleConfigureOpenChange,
-            configurePopoverContent,
-            onRefinePrompt,
-            updatePromptRootField,
-            siblingGroups,
-            isPresentSiblingGroup,
-            codeRuntime,
-            handleRuntimeChange,
-        ],
-    )
-
-    // ========== FIELD CONTENT SLOT ==========
-    const fieldContentSlot = useCallback(
-        (props: FieldContentSlotProps) => {
-            const isTopLevel = props.path.length === 1
-            if (!isTopLevel) return props.defaultRender()
-
-            const fieldKey = String(props.field.key)
-
-            // Hide llms field when the model popover handles it
-            // (canonical schema — llms[0] is shown in the model popover on the messages header)
-            if (fieldKey === "llms" && promptModelInfo?.isRootLevel) {
-                return null
-            }
-
-            // Sibling group (hook/code): render the dedicated control, not the
-            // schema renderer.
-            const isSibling = isPresentSiblingGroup(fieldKey)
-            const isCollapsed = !!collapsedSections[fieldKey]
-
-            if (isSibling) {
-                return (
-                    <HeightCollapse open={!isCollapsed}>
-                        <div className="px-4 py-3 flex flex-col gap-3">
-                            {fieldKey === "schemas" ? (
-                                <SchemasConfigControl
-                                    value={siblingGroups[fieldKey] as Record<string, unknown>}
-                                    onChange={(next) => props.onChange(next)}
-                                    disabled={disabled}
-                                />
-                            ) : fieldKey === "hook" ? (
-                                <HookConfigControl
-                                    value={siblingGroups[fieldKey] as Record<string, unknown>}
-                                    onChange={(next) => props.onChange(next)}
-                                    disabled={disabled}
-                                />
-                            ) : (
-                                <CodeConfigControl
-                                    value={siblingGroups[fieldKey] as Record<string, unknown>}
-                                    onChange={(next) => props.onChange(next)}
-                                    disabled={disabled}
-                                />
-                            )}
-                        </div>
-                    </HeightCollapse>
-                )
-            }
-
-            // Scalar/array params render directly, without HeightCollapse.
-            const fieldValue = parameters[fieldKey]
-            if (
-                fieldValue === null ||
-                fieldValue === undefined ||
-                typeof fieldValue !== "object" ||
-                Array.isArray(fieldValue)
-            ) {
-                return <div className="px-4 py-1.5">{props.defaultRender()}</div>
-            }
-
-            // The agent body is always expanded (its header is suppressed above), so it renders
-            // without the HeightCollapse toggle. Detected via the same marker as the header
-            // suppression — both the new `agent-template` and legacy `agent_config` name, so it
-            // holds across the rename (checking only `agent_config` here sent modern agent fields
-            // to the generic else branch, re-adding the collapse and the wider py-3 inset).
-            const fieldSchema = schema?.properties
-                ? (schema.properties as Record<string, Record<string, unknown>>)[fieldKey]
-                : null
-            const isAgentMarker = (v: unknown) => v === "agent-template" || v === "agent_config"
-            if (
-                isAgentMarker(fieldSchema?.["x-ag-type-ref"]) ||
-                isAgentMarker(fieldSchema?.["x-ag-type"])
-            ) {
-                // pt-0 (not py-3) tightens the gap between the region header and the first section
-                // row; pb-3 keeps the bottom breathing room. The loading fallback below matches.
-                return <div className="px-4 pb-3 pt-0">{props.defaultRender()}</div>
-            }
-
-            return (
-                <HeightCollapse open={!isCollapsed}>
-                    <div className="px-4 py-3">{props.defaultRender()}</div>
-                </HeightCollapse>
-            )
-        },
-        [
-            collapsedSections,
-            parameters,
-            schema,
-            siblingGroups,
-            isPresentSiblingGroup,
-            disabled,
-            promptModelInfo?.isRootLevel,
-        ],
-    )
-
-    // Stable slots object — an inline literal would churn the drill-in provider's context value.
-    const drillInSlots = useMemo(
-        () => ({
-            fieldHeader: fieldHeaderSlot,
-            fieldActions: fieldActionsSlot,
-            fieldContent: fieldContentSlot,
-        }),
-        [fieldHeaderSlot, fieldActionsSlot, fieldContentSlot],
-    )
+    const {
+        configurePopoverContent,
+        handleConfigureOpenChange,
+        isModelConfigOpen,
+        promptModelInfo,
+        updatePromptRootField,
+    } = useModelConfigurePopover({
+        activeData,
+        disabled,
+        dispatchUpdate,
+        llmProviderConfig,
+        parameters,
+        revisionId,
+        schema,
+        serverData,
+    })
+
+    // ========== DRILL-IN FIELD SLOTS ==========
+    const drillInSlots = useFieldSlots({
+        activeData,
+        codeRuntime,
+        collapsedSections,
+        configurePopoverContent,
+        disabled,
+        feedbackMode,
+        handleConfigureOpenChange,
+        handleRuntimeChange,
+        isModelConfigOpen,
+        isPresentSiblingGroup,
+        onRefinePrompt,
+        parameters,
+        promptModelInfo,
+        revisionId,
+        schema,
+        setFeedbackMode,
+        siblingGroups,
+        stickyHeaderTop,
+        toggleSection,
+        updatePromptRootField,
+    })
 
     // ========== LOADING / EMPTY STATE ==========
     const isConfigLoading = schemaQuery.isPending && !hasRenderableConfigSections(activeData)
